requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy==1.18.5
pandas
scikit-learn
tensorflow
joblib

# nlp
nltk
gensim==4.0.1

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
uvicorn
fastapi

# utilities
six>=1.14
memoized-property
termcolor
python-dotenv
